Task2/requirements.txt
absl-py==1.4.0
alabaster==0.7.13
astunparse==1.6.3
attrs==23.1.0
Babel==2.12.1
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
click==7.1.2
dill==0.3.6
docutils==0.18.1
flatbuffers==23.5.26
future==0.18.3
gast==0.4.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.59.1
grpcio==1.56.0
h5py==3.9.0
idna==3.4
imagesize==1.4.1
indic-nlp-library==0.92
Jinja2==3.1.2
keras==2.13.1
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.3
Morfessor==2.0.6
numpy==1.24.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.3
promise==2.3
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.15.1
python-dateutil==2.8.2
pytz==2023.3
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
snowballstemmer==2.2.0
Sphinx==6.2.1
sphinx-argparse==0.4.0
sphinx-rtd-theme==1.2.2
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-addons==0.21.0
tensorflow-datasets==3.2.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.32.0
tensorflow-metadata==1.13.1
termcolor==2.3.0
tf2crf==0.1.33
tqdm==4.65.0
typeguard==2.13.3
typing_extensions==4.5.0
tzdata==2023.3
urduhack==1.1.1
urllib3==1.26.16
Werkzeug==2.3.6
wrapt==1.15.0
